Get the error handler in place before starting to load sounds.
Otherwise, nothing's listening for the IOErrorEvent yet and things blow up
spectacularly if the file isn't there.

I'm a little freaked out that it's still going straight into calling onInit()
in case of a streaming sound, but hopefully anything using that is behaving
properly in the face of a streaming error.